+++ src/plejdService.ts
@@ -1,38 +1,38 @@
-import { Logger } from 'homebridge';
-import { UserInputConfig } from './model/userInputConfig.js';
+import { Logger } from "homebridge";
+import { UserInputConfig } from "./model/userInputConfig.js";
 import {
   plejdChalResp as plejdCharResp,
   plejdEncodeDecode,
   reverseBuffer,
-} from './plejdUtils.js';
+} from "./plejdUtils.js";
 
-import { randomBytes } from 'crypto';
-import noble from '@abandonware/noble';
-import { PLEJD_PING_TIMEOUT, PLEJD_WRITE_TIMEOUT } from './settings.js';
-import { delay } from './utils.js';
+import { randomBytes } from "crypto";
+import noble from "@abandonware/noble";
+import { PLEJD_PING_TIMEOUT, PLEJD_WRITE_TIMEOUT } from "./settings.js";
+import { delay } from "./utils.js";
 
 /**
  * Plejd BLE UUIDs
  */
 enum PlejdCharacteristics {
-  Service = '31ba000160854726be45040c957391b5',
-  LightLevel = '31ba000360854726be45040c957391b5',
-  Data = '31ba000460854726be45040c957391b5',
-  LastData = '31ba000560854726be45040c957391b5',
-  Auth = '31ba000960854726be45040c957391b5',
-  Ping = '31ba000a60854726be45040c957391b5',
+  Service = "31ba000160854726be45040c957391b5",
+  LightLevel = "31ba000360854726be45040c957391b5",
+  Data = "31ba000460854726be45040c957391b5",
+  LastData = "31ba000560854726be45040c957391b5",
+  Auth = "31ba000960854726be45040c957391b5",
+  Ping = "31ba000a60854726be45040c957391b5",
 }
 
 enum PlejdCommand {
-  OnOffState = '0097',
-  StateBrightness = '00c8',
-  Brightness = '0098', // 0-255
-  Scene = '0021',
-  Time = '001b',
-  RequestResponse = '0102',
-  RequestNoResponse = '0110',
-  RequestReadValue = '0103',
-  ButtonClick = '0006',
+  OnOffState = "0097",
+  StateBrightness = "00c8",
+  Brightness = "0098", // 0-255
+  Scene = "0021",
+  Time = "001b",
+  RequestResponse = "0102",
+  RequestNoResponse = "0110",
+  RequestReadValue = "0103",
+  ButtonClick = "0006",
 }
 
 export class PlejdService {
@@ -70,18 +70,18 @@ export class PlejdService {
 
     const payload = Buffer.from(
       !brightness || brightness === 0
-        ? identifier.toString(16).padStart(2, '0') +
+        ? identifier.toString(16).padStart(2, "0") +
             PlejdCommand.RequestNoResponse +
             PlejdCommand.OnOffState +
-            (isOn ? '01' : '00')
-        : identifier.toString(16).padStart(2, '0') +
+            (isOn ? "01" : "00")
+        : identifier.toString(16).padStart(2, "0") +
             PlejdCommand.RequestNoResponse +
             PlejdCommand.Brightness +
-            '01' +
+            "01" +
             Math.round(2.55 * brightness)
               .toString(16)
-              .padStart(4, '0'),
-      'hex',
+              .padStart(4, "0"),
+      "hex",
     );
 
     const data = plejdEncodeDecode(
@@ -94,20 +94,20 @@ export class PlejdService {
   };
 
   configureBLE = () => {
-    noble.on('stateChange', async (state) => {
+    noble.on("stateChange", async (state) => {
       this.log.debug(`Noble State changed: ${state}`);
-      if (state === 'poweredOn') {
-        this.log.info('Scanning for Plejd devices as we started...');
+      if (state === "poweredOn") {
+        this.log.info("Scanning for Plejd devices as we started...");
         await noble.startScanningAsync([PlejdCharacteristics.Service], false);
       }
     });
 
-    noble.on('warning', (msg: string) =>
-      this.log.warn('Noble warning: ', msg),
-    );
+    noble.on("warning", (msg: string) => {
+      this.log.warn("Noble warning: ", msg);
+    });
 
     noble.on(
-      'discover',
+      "discover",
       async (peripheral) => await this.onDiscover(peripheral),
     );
   };
@@ -122,28 +122,35 @@ export class PlejdService {
 
     try {
       await peripheral.connectAsync();
-    } catch (error) {
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    } catch (error: any) {
+      if (error.toString().includes("Peripheral already connected")) {
+        await peripheral?.disconnectAsync();
+        await this.onDiscover(peripheral);
+        return;
+      }
+
       this.log.error(
         `Connecting failed | ${peripheral.advertisement.localName} | addr: ${peripheral.address}) - err: ${error}`,
       );
       return;
     }
 
-    this.connectedPeripheral = peripheral;
-
-    peripheral.once('disconnect', async () => {
-      this.log.info('Disconnected from mesh');
+    this.connectedPeripheral = peripheral.once("disconnect", async () => {
+      this.log.info("Disconnected from mesh");
       noble.reset();
-      if (noble._state === 'poweredOn') {
-        this.log.info('Scanning for Plejd devices as we are disconnected from mesh...');
+      if (noble._state === "poweredOn") {
+        this.log.info(
+          "Scanning for Plejd devices as we are disconnected from mesh...",
+        );
         await noble.startScanningAsync([PlejdCharacteristics.Service], false);
       }
     });
 
     const characteristics = await this.discoverCaracteristics(peripheral);
     if (!characteristics) {
-      this.log.error('Failed to discover characteristics, disconnecting...');
-      if (peripheral.state === 'connected') {
+      this.log.error("Failed to discover characteristics, disconnecting...");
+      if (peripheral.state === "connected") {
         await peripheral.disconnectAsync();
       }
       return;
@@ -162,7 +169,7 @@ export class PlejdService {
     );
 
     this.addressBuffer = reverseBuffer(
-      Buffer.from(String(addr).replace(/:/g, ''), 'hex'),
+      Buffer.from(String(addr).replace(/:/g, ""), "hex"),
     );
 
     const services = [PlejdCharacteristics.Service];
@@ -208,7 +215,7 @@ export class PlejdService {
 
     if (!authChar || !lastDataChar || !pingChar) {
       this.log.error(
-        'Unable to extract characteristic during discovery',
+        "Unable to extract characteristic during discovery",
         authChar,
         lastDataChar,
         pingChar,
@@ -237,12 +244,12 @@ export class PlejdService {
         return;
       }
       this.log.debug(
-        `BLE command sent to ${this.addressBuffer?.toString('hex') ?? 'Unknown'} | ${data.length} bytes | ${data.toString('hex')}`,
+        `BLE command sent to ${this.addressBuffer?.toString("hex") ?? "Unknown"} | ${data.length} bytes | ${data.toString("hex")}`,
       );
       try {
         await this.dataCharacteristic.writeAsync(data, false);
       } catch (error) {
-        this.log.error('Failed to send data to device, will retry: ', error);
+        this.log.error("Failed to send data to device, will retry: ", error);
         this.sendQueue.unshift(data);
         return;
       }
@@ -258,13 +265,13 @@ export class PlejdService {
         const pong = await pingChar.readAsync();
         if (((ping[0] + 1) & 0xff) !== pong[0]) {
           this.log.error(
-            'Ping pong communication failed, missing pong response',
+            "Ping pong communication failed, missing pong response",
           );
         }
         await delay(PLEJD_PING_TIMEOUT);
       } catch (error) {
         this.log.warn(
-          'Ping failed, device disconnected, will retry to connect to mesh: ',
+          "Ping failed, device disconnected, will retry to connect to mesh: ",
           error,
         );
       }
@@ -275,7 +282,11 @@ export class PlejdService {
     data: Buffer,
     isNotification: boolean,
   ) => {
-    if (!this.connected() || !this.addressBuffer || this.addressBuffer?.byteLength === 0) {
+    if (
+      !this.connected() ||
+      !this.addressBuffer ||
+      this.addressBuffer?.byteLength === 0
+    ) {
       return;
     }
 
@@ -285,13 +296,13 @@ export class PlejdService {
       data,
     );
     const id = parseInt(decodedData[0].toString(), 10);
-    const command = decodedData.toString('hex', 3, 5);
+    const command = decodedData.toString("hex", 3, 5);
 
-    const isOn = parseInt(decodedData.toString('hex', 5, 6), 10) === 1;
+    const isOn = parseInt(decodedData.toString("hex", 5, 6), 10) === 1;
 
     const commandType =
       Object.values(PlejdCommand).find((x) => x.toString() === command) ??
-      'Unknown';
+      "Unknown";
 
     const d = {
       id: id,
@@ -299,23 +310,23 @@ export class PlejdService {
       commandType: commandType,
       on: isOn,
       Notification: isNotification,
-      payload: decodedData.toString('hex'),
+      payload: decodedData.toString("hex"),
     };
-    this.log.debug('Handle BLE notification', d);
+    this.log.debug("Handle BLE notification", d);
 
     switch (command) {
       case PlejdCommand.Time: {
         const arg = parseInt(
-          reverseBuffer(decodedData.subarray(5, 9)).toString('hex'),
+          reverseBuffer(decodedData.subarray(5, 9)).toString("hex"),
           16,
         );
         const date = new Date(arg * 1000);
-        this.log.debug('Time sync: ' + date.toString());
+        this.log.debug("Time sync: " + date.toString());
         break;
       }
       case PlejdCommand.Brightness:
       case PlejdCommand.StateBrightness: {
-        const dim = parseInt(decodedData.toString('hex', 7, 8), 16);
+        const dim = parseInt(decodedData.toString("hex", 7, 8), 16);
 
         // Convert to Homebridge 1-100
         const converted = dim === 0 ? 1 : (100 / 255) * dim;
@@ -335,7 +346,7 @@ export class PlejdService {
         this.onUpdate(id, isOn);
         this.log.warn(
           `Unknown | command: ${command} | id: ${id} | ${decodedData.toString(
-            'hex',
+            "hex",
           )}`,
         );
       }
@@ -348,7 +359,7 @@ export class PlejdService {
   ) => {
     this.startPlejdPing(pingChar);
     await lastDataChar.subscribeAsync();
-    lastDataChar.on('data', async (data, isNotification) => {
+    lastDataChar.on("data", async (data, isNotification) => {
       if (!this.connected()) {
         await lastDataChar.unsubscribeAsync();
         return;
@@ -366,5 +377,6 @@ export class PlejdService {
     );
   };
 
-  private connected = () => this.connectedPeripheral?.state === 'connected';
+  private connected = () =>
+    this.connectedPeripheral && this.connectedPeripheral.state === "connected";
 }